// src/API.ts
export type Blog = {
  __typename: "Blog",
  id: string,
  name: string,
  body?: string | null,
  createdAt: string,
  updatedAt: string,
};
// ...
export type ModelBlogConnection = {
  __typename: "ModelBlogConnection",
  items:  Array<Blog | null >,
  nextToken?: string | null,
};
// ...
export type ListBlogsQuery = {
  listBlogs?:  {
    __typename: "ModelBlogConnection",
    items:  Array< {
      __typename: "Blog",
      id: string,
      name: string,
      body?: string | null,
      createdAt: string,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};